YOLO+PDLPR/PDLPR/requirements.txt
torch>=1.12
torchvision>=0.13
Pillow>=8.0
numpy
tqdm
